+++ rockpool/devices/xylo/syns65302/reg.py
@@ -1,3 +1,7 @@
+"""
+Defines register names for Xylo™Audio 3 SYNS65302
+"""
+
 # ===========================================================================
 # Description     : test file
 # Author          : Bai, Xin