+++ tests/tests_default/test_losses.py
@@ -2,6 +2,8 @@ import pytest
 
 
 def test_peak_loss():
+
+    pytest.importorskip("torch")
     import torch
     from rockpool.nn.losses import PeakLoss
 
@@ -67,6 +69,7 @@ def test_peak_loss():
 
 
 def test_binary_peak_loss():
+    pytest.importorskip("torch")
     import torch
     from rockpool.nn.losses import BinaryPeakLoss
 
@@ -137,6 +140,7 @@ def test_binary_peak_loss():
 
 
 def test_mse_loss():
+    pytest.importorskip("torch")
     import torch
     from rockpool.nn.losses import MSELoss